Fix scripts.

+++ Generator/run_tests.sh
@@ -1,8 +1,9 @@
 #!/usr/bin/env bash
-if [ "$1" != "--no-build" ]; then
+FILE_NAME="cuckoo_generator.app"
+if [[ "$1" != "--no-build" && -f "$FILE_NAME"]]; then
 	rm -rf Build
 	mkdir Build
-	xcodebuild -project 'CuckooGenerator.xcodeproj' -scheme 'CuckooGenerator' -configuration 'Release' CONFIGURATION_BUILD_DIR=$(pwd)/Build clean build
+	xcodebuild -project 'CuckooGenerator.xcodeproj' -scheme 'CuckooGenerator' -configuration 'Release' CONFIGURATION_BUILD_DIR="$PWD/Build" clean build
 fi
 cd Tests
 cucumber
